// nodejs/src/cdp/types.ts
export type HogBytecode = any[]
// ...
export type HogFunctionMasking = {
    ttl: number | null
    hash: string
    bytecode: HogBytecode
    threshold: number | null
}

export interface HogFunctionFilters {
    source?: 'events' | 'person-updates' | 'data-warehouse-table' // Special case to identify what kind of thing this filters on
    events?: HogFunctionFilterEvent[]
    actions?: HogFunctionFilterAction[]
    properties?: Record<string, any>[] // Global property filters that apply to all events
    filter_test_accounts?: boolean
    bytecode?: HogBytecode
}
// ...
export type HogFunctionInvocationGlobalsWithInputs = HogFunctionInvocationGlobals & {
    inputs: Record<string, any>
}
// ...
export interface HogFunctionTiming {
    kind: 'hog' | 'async_function'
    duration_ms: number
}

// IMPORTANT: All queue names should be lowercase and only [A-Z0-9] characters are allowed.
export const CYCLOTRON_INVOCATION_JOB_QUEUES = ['hog', 'hogoverflow', 'hogflow', 'email'] as const
export type CyclotronJobQueueKind = (typeof CYCLOTRON_INVOCATION_JOB_QUEUES)[number]

export const CYCLOTRON_JOB_QUEUE_SOURCES = ['postgres', 'postgres-v2', 'kafka'] as const
export type CyclotronJobQueueSource = (typeof CYCLOTRON_JOB_QUEUE_SOURCES)[number]

// Agnostic job invocation type
export type CyclotronJobInvocation = {
    id: string
    teamId: Team['id']
    functionId: string
    // Optional parent run ID, e.g. if this invocation is part of a batch workflow run
    parentRunId?: string | null
    state: Record<string, any> | null
    // The queue that the invocation is on
    queue: CyclotronJobQueueKind
    // Optional parameters for that queue to use
    queueParameters?: CyclotronInvocationQueueParametersType | null
    // Priority of the invocation
    queuePriority: number
    // When the invocation is scheduled to run
    queueScheduledAt?: DateTime
    // Metadata for the invocation - TODO: check when this gets cleared
    queueMetadata?: Record<string, any> | null
    // Where the invocation came from (kafka or postgres)
    queueSource?: CyclotronJobQueueSource
}
// ...
export type CyclotronJobInvocationHogFunctionContext = {
    globals: HogFunctionInvocationGlobalsWithInputs
    vmState?: VMState
    timings: HogFunctionTiming[]
    attempts: number // Indicates the number of times this invocation has been attempted (for example if it gets scheduled for retries)
    actionId?: string // The hogflow action node ID, used for metrics instance_id when executing within a workflow
}

export type CyclotronJobInvocationHogFunction = CyclotronJobInvocation & {
    state: CyclotronJobInvocationHogFunctionContext
    hogFunction: HogFunctionType
}
// ...
export type HogFunctionInputSchemaType = {
    type:
        | 'string'
        | 'number'
        | 'boolean'
        | 'dictionary'
        | 'choice'
        | 'json'
        | 'integration'
        | 'integration_field'
        | 'email'
        | 'native_email'
        | 'posthog_assignee'
        | 'posthog_ticket_tags'
    key: string
    label?: string
    choices?: { value: string; label: string }[]
    required?: boolean
    default?: any
    secret?: boolean
    hidden?: boolean
    description?: string
    integration?: string
    integration_key?: string
    requires_field?: string
    integration_field?: string
    requiredScopes?: string
    /**
     * templating: true indicates the field supports templating. Alternatively
     * it can be set to 'hog' or 'liquid' to specify the default templating engine to use.
     */
    templating?: boolean | 'hog' | 'liquid'
}

export type HogFunctionTypeType =
    | 'destination'
    | 'transformation'
    | 'internal_destination'
    | 'source_webhook'
    | 'warehouse_source_webhook'
    | 'site_destination'

export interface HogFunctionMappingType {
    inputs_schema?: HogFunctionInputSchemaType[]
    inputs?: Record<string, CyclotronInputType> | null
    filters?: HogFunctionFilters | null
}

export type HogFunctionType = {
    id: string
    type: HogFunctionTypeType
    team_id: number
    name: string
    enabled: boolean
    deleted: boolean
    hog: string
    bytecode: HogBytecode
    inputs_schema?: HogFunctionInputSchemaType[]
    inputs?: Record<string, CyclotronInputType | null>
    encrypted_inputs?: Record<string, CyclotronInputType>
    filters?: HogFunctionFilters | null
    mappings?: HogFunctionMappingType[] | null
    masking?: HogFunctionMasking | null
    template_id?: string
    execution_order?: number
    created_at: string
    updated_at: string
    metadata?: Record<string, any>
    batch_export_id?: string | null
}